Move default and expected value NRQL assertions into a separate file
